+++ src/main/java/com/example/learnSpringBoot/interceptor/PersonInterceptorAppConfig.java
@@ -1,0 +1,22 @@
+package com.example.learnSpringBoot.interceptor;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.stereotype.Component;
+import org.springframework.web.servlet.config.annotation.EnableWebMvc;
+import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+
+@EnableWebMvc
+@Configuration
+public class PersonInterceptorAppConfig implements WebMvcConfigurer{
+	
+	@Autowired
+	PersonInterceptor personInterceptor;
+	
+	@Override
+	public void addInterceptors(InterceptorRegistry registry) {
+		registry.addInterceptor(personInterceptor);
+	}
+
+}