+++ changeLog.md
@@ -5,3 +5,6 @@
 
 #Commit 4 [15-03-21]
 - Exception Handling using @ControllerAdvice and @ExceptionHandler
+
+#Commit 5 [19-02-21]
+- Interceptor implemented. As the name suggests intercepts the request/response.